バースト時に敵が吹き飛ぶように（GuardBurst少し変更しました）

+++ Assets/Script/OguraScript/StatusComponent.cs
@@ -1,3 +1,12 @@
+//======================================================================
+// StatusComponent.cs
+//======================================================================
+// 開発履歴
+//
+// 2022/03/05 製作開始 ステータス関連
+//
+//======================================================================
+
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
@@ -7,7 +16,7 @@ using UnityEngine;
 public class StatusComponent : MonoBehaviour
 {
     public int Level { get; set; }
-    public float HP { get; set; }
-    public float Attack { get; set; }
+    public int HP { get; set; }
+    public int Attack { get; set; }
     public float Speed { get; set; }
 }
